# openapi.yml
openapi: 3.0.3
info:
  title: Blog API
  version: 1.0.0
  description: Blog API with authentication, posts and comments.
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /posts:
    get:
      tags: [Post]
      summary: List posts
      description: Returns a paginated list of posts, newest first. Public.
      operationId: listPosts
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          $ref: '#/components/responses/PostListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Post]
      summary: Create a post
      description: Creates a new post owned by the authenticated user.
      operationId: createPost
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          $ref: '#/components/responses/PostSingleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /posts/{postId}:
    parameters:
      - $ref: '#/components/parameters/PostIdParam'

    get:
      tags: [Post]
      summary: Get a post by ID
      description: Returns a single post. Public.
      operationId: getPostById
      security: []
      responses:
        '200':
          $ref: '#/components/responses/PostSingleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Post]
      summary: Update a post
      description: Partially updates a post. Only the post's owner or an admin may update it.
      operationId: updatePost
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          $ref: '#/components/responses/PostSingleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Post]
      summary: Delete a post
      description: Deletes a post and all of its comments. Only the post's owner or an admin may delete it.
      operationId: deletePost
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/PostDeleteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /posts/{postId}/comments:
    parameters:
      - $ref: '#/components/parameters/PostIdParam'

    get:
      tags: [Comment]
      summary: List comments on a post
      description: Returns a paginated list of comments for the given post, oldest first. Public.
      operationId: listPostComments
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          $ref: '#/components/responses/CommentListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Comment]
      summary: Add a comment to a post
      description: Creates a new comment on the given post, owned by the authenticated user.
      operationId: createComment
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          $ref: '#/components/responses/CommentSingleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /comments/{commentId}:
    parameters:
      - $ref: '#/components/parameters/CommentIdParam'

    patch:
      tags: [Comment]
      summary: Update a comment
      description: Partially updates a comment. Only the comment's owner may update it.
      operationId: updateComment
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentUpdate'
      responses:
        '200':
          $ref: '#/components/responses/CommentSingleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Comment]
      summary: Delete a comment
      description: Deletes a comment. Only the comment's owner or an admin may delete it.
      operationId: deleteComment
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/CommentDeleteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accesstoken
      description: JWT access token set as an httpOnly cookie
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refreshtoken
      description: JWT refresh token set as an httpOnly cookie, used only on the refresh endpoint
    googleOAuth:
      type: oauth2
      description: Google OAuth 2.0 login flow. The client is redirected to /auth/google, which forwards to Google, then back to /auth/google/callback.
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID scope
            email: Grants access to the user's email address
            profile: Grants access to the user's basic profile info
  parameters:
    PostIdParam:
      name: postId
      in: path
      required: true
      description: UUID of the post
      schema:
        type: string
        format: uuid
    CommentIdParam:
      name: commentId
      in: path
      required: true
      description: UUID of the comment
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      required: false
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        status:
          type: string
          enum: [fail, error]
          description: "'fail' for operational/validation errors, 'error' for unexpected server errors"
        message:
          type: string
          example: Invalid Input
        requestId:
          type: string
          format: uuid
      required: [success, status, message]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
          example: User logged in successfully
        meta:
          type: object
      required: [success, data]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3

    # ----- Auth -----
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        title:
          type: string
          example: My first blog post
        content:
          type: string
          example: Hello world, this is the body of my first post.
        authorId:
          type: string
          format: uuid
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - authorId
        - createdAt
        - updatedAt
    PostCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: My first blog post
        content:
          type: string
          minLength: 1
          example: Hello world, this is the body of my first post.
      required:
        - title
        - content
    PostUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          example: My updated blog post
        content:
          type: string
          minLength: 1
          example: Updated body of the post.
      minProperties: 1
      description: At least one field must be provided.
    PostSuccessResponse:
      description: Envelope wrapping a single post in `data`.
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Post'
          required: [data]
    PostListSuccessResponse:
      description: Envelope wrapping an array of posts in `data`, with pagination in `meta`.
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Post'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
          required: [data]

    # ----- Comment -----
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 9b2f1c3e-4d5a-4e6f-8a7b-1c2d3e4f5a6b
        content:
          type: string
          example: Great post, thanks for sharing!
        postId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        authorId:
          type: string
          format: uuid
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - postId
        - authorId
        - createdAt
        - updatedAt
    CommentCreate:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          example: Great post, thanks for sharing!
      required:
        - content
    CommentUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          example: Great post, thanks for sharing! (edited)
      minProperties: 1
      description: At least one field must be provided.
    CommentSuccessResponse:
      description: Envelope wrapping a single comment in `data`.
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Comment'
          required: [data]
    CommentListSuccessResponse:
      description: Envelope wrapping an array of comments in `data`, with pagination in `meta`.
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
          required: [data]
  responses:
    PostSingleResponse:
      description: A single post.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PostSuccessResponse'
    PostListResponse:
      description: A paginated list of posts.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PostListSuccessResponse'
    PostDeleteResponse:
      description: Post deleted successfully.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessResponse'
              - type: object
                properties:
                  data:
                    type: object
                    nullable: true
                    example: null
                  message:
                    type: string
                    example: Post deleted successfully
    CommentSingleResponse:
      description: A single comment.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommentSuccessResponse'
    CommentListResponse:
      description: A paginated list of comments.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommentListSuccessResponse'
    CommentDeleteResponse:
      description: Comment deleted successfully.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/SuccessResponse'
              - type: object
                properties:
                  data:
                    type: object
                    nullable: true
                    example: null
                  message:
                    type: string
                    example: Comment deleted successfully
    ValidationError:
      description: The request body or parameters failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: fail
            message: Invalid Input
            requestId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    UnauthorizedError:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: fail
            message: You are not logged in. Please log in to get access.
            requestId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    ForbiddenError:
      description: The authenticated user does not have permission to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: fail
            message: You do not have permission to perform this action.
            requestId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    NotFoundError:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: fail
            message: No resource found with that ID.
            requestId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
...
    InternalServerError:
      description: An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            status: error
            message: Something went wrong.
            requestId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
security:
  - cookieAuth: []